requirements.txt
alembic==1.13.2
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1716290248287/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/anyio_1717693030552/work
bcrypt==4.0.1
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
cffi==1.17.1
click @ file:///D:/bld/click_1692312014553/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
cryptography==43.0.1
DateTime==5.5
dnspython @ file:///home/conda/feedstock_root/build_artifacts/dnspython_1709190728464/work
ecdsa==0.19.0
email_validator @ file:///home/conda/feedstock_root/build_artifacts/email-validator-meta_1718984626658/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1720869315914/work
fastapi @ file:///home/conda/feedstock_root/build_artifacts/fastapi_1725577067245/work
fastapi-cli @ file:///home/conda/feedstock_root/build_artifacts/fastapi-cli_1722597621840/work
greenlet==3.1.0
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1664132893548/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
hpack==4.0.0
httpcore @ file:///home/conda/feedstock_root/build_artifacts/httpcore_1711596990900/work
httpx @ file:///home/conda/feedstock_root/build_artifacts/httpx_1724778349782/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1724450538981/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1715127149914/work
Mako==1.3.5
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1686175045316/work
MarkupSafe @ file:///D:/bld/markupsafe_1724959577091/work
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1704317613764/work
passlib==1.7.4
psycopg==3.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic @ file:///home/conda/feedstock_root/build_artifacts/pydantic_1720293063581/work
pydantic-settings==2.5.2
pydantic_core @ file:///D:/bld/pydantic-core_1720041343998/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart @ file:///home/conda/feedstock_root/build_artifacts/python-multipart_1707760088566/work
pytz==2024.2
rich @ file:///home/conda/feedstock_root/build_artifacts/rich-split_1709150387247/work/dist
rsa==4.9
setuptools==73.0.1
shellingham @ file:///home/conda/feedstock_root/build_artifacts/shellingham_1698144360966/work
six==1.16.0
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1708952932303/work
SQLAlchemy==2.0.35
starlette @ file:///home/conda/feedstock_root/build_artifacts/starlette-recipe_1725309582791/work
typer==0.12.5
typer-slim==0.12.5
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.1
uvicorn @ file:///D:/bld/uvicorn-split_1723660175812/work
wheel==0.44.0
zope.interface==7.0.3
